# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PCRE API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/feedback": {
      "post": {
        "tags": [
          "insights"
        ],
        "summary": "Submit Feedback",
        "operationId": "submit_feedback_api_v1_feedback_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedbackIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatedOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/pilot/summary": {
      "get": {
        "tags": [
          "insights"
        ],
        "summary": "Pilot Summary",
        "operationId": "pilot_summary_api_v1_admin_pilot_summary_get",
        "parameters": [
          {
            "name": "days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 28,
              "minimum": 1,
              "default": 7,
              "title": "Days"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PilotSummaryOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/errors": {
      "get": {
        "tags": [
          "insights"
        ],
        "summary": "Errors",
        "operationId": "errors_api_v1_admin_errors_get",
        "parameters": [
          {
            "name": "days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 30,
              "minimum": 1,
              "default": 30,
              "title": "Days"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ErrorRowOut"
                  },
                  "title": "Response Errors Api V1 Admin Errors Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ActiveDayOut": {
        "properties": {
          "day": {
            "type": "string",
            "title": "Day"
          },
          "students": {
            "type": "integer",
            "title": "Students"
          }
        },
        "type": "object",
        "required": [
          "day",
          "students"
        ],
        "title": "ActiveDayOut"
      },
...
      "CreatedOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "id"
        ],
        "title": "CreatedOut"
      },
...
      "ErrorRowOut": {
        "properties": {
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "route": {
            "type": "string",
            "title": "Route"
          },
          "status_code": {
            "type": "integer",
            "title": "Status Code"
          },
          "error_code": {
            "type": "string",
            "title": "Error Code"
          },
          "exception_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Exception Type"
          },
          "occurred_at": {
            "type": "string",
            "title": "Occurred At"
          }
        },
        "type": "object",
        "required": [
          "request_id",
          "route",
          "status_code",
          "error_code",
          "exception_type",
          "occurred_at"
        ],
        "title": "ErrorRowOut"
      },
      "FeedbackIn": {
        "properties": {
          "context_type": {
            "type": "string",
            "enum": [
              "lesson",
              "general"
            ],
            "title": "Context Type"
          },
          "context_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context Id"
          },
          "rating": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 5.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Rating"
          },
          "message": {
            "type": "string",
            "maxLength": 1000,
            "title": "Message",
            "default": ""
          },
          "page": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Page"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "context_type"
        ],
        "title": "FeedbackIn",
        "description": "Valoración de una lección o escenario (1–5, comentario opcional) o comentario\ngeneral con la página como contexto (REQ-15)."
      },
...
      "FeedbackRowOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "context_type": {
            "type": "string",
            "title": "Context Type"
          },
          "context_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context Id"
          },
          "rating": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rating"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "page": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "internal": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Internal"
          }
        },
        "type": "object",
        "required": [
          "id",
          "email",
          "context_type",
          "context_id",
          "rating",
          "message",
          "page",
          "created_at"
        ],
        "title": "FeedbackRowOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PilotSummaryOut": {
        "properties": {
          "days": {
            "type": "integer",
            "title": "Days"
          },
          "internal_excluded": {
            "type": "boolean",
            "title": "Internal Excluded"
          },
          "active_by_day": {
            "items": {
              "$ref": "#/components/schemas/ActiveDayOut"
            },
            "type": "array",
            "title": "Active By Day"
          },
          "attempts": {
            "type": "integer",
            "title": "Attempts"
          },
          "lessons_completed": {
            "type": "integer",
            "title": "Lessons Completed"
          },
          "reviews_done": {
            "type": "integer",
            "title": "Reviews Done"
          },
          "reviews_due": {
            "type": "integer",
            "title": "Reviews Due"
          },
          "diagnostics": {
            "type": "integer",
            "title": "Diagnostics"
          },
          "checkpoints": {
            "type": "integer",
            "title": "Checkpoints"
          },
          "average_rating": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Average Rating"
          },
          "ratings": {
            "type": "integer",
            "title": "Ratings"
          },
          "latest_comments": {
            "items": {
              "$ref": "#/components/schemas/FeedbackRowOut"
            },
            "type": "array",
            "title": "Latest Comments"
          },
          "open_reports": {
            "type": "integer",
            "title": "Open Reports"
          },
          "server_errors": {
            "type": "integer",
            "title": "Server Errors"
          },
          "latest_errors": {
            "items": {
              "$ref": "#/components/schemas/ErrorRowOut"
            },
            "type": "array",
            "title": "Latest Errors"
          }
        },
        "type": "object",
        "required": [
          "days",
          "internal_excluded",
          "active_by_day",
          "attempts",
          "lessons_completed",
          "reviews_done",
          "reviews_due",
          "diagnostics",
          "checkpoints",
          "average_rating",
          "ratings",
          "latest_comments",
          "open_reports",
          "server_errors",
          "latest_errors"
        ],
        "title": "PilotSummaryOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}